requirements.txt
fastapi==0.115.6
jose==1.0.0
passlib==1.7.4
pydantic==2.10.5
python-dotenv==1.0.1
python-jose==3.3.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2

requests~=2.32.3
alembic~=1.15.2